% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwr.R
\name{mcwr_exampledata}
\alias{mcwr_exampledata}
\title{Load mcwr example data}
\usage{
mcwr_exampledata(snipnum)
}
\arguments{
\item{snipnum}{a integer [1-9]. Determines the data frame returned.}
}
\value{
A data.frame, suitable for further processing by \code{mcwr_*()} functions.
}
\description{
This function is used in examples sections and loads mcwr example data.
It only exists for documentation purposes.
}
